Adds force_update flag to MappedObject with methods to set the flag and return the flag
Add check on results return in QueueHandler to avoid error of calling $results->records on non-
existant object.
Updates RestClient::apiCall to throw RestExceptions in all error cases
Updates PullBase to detect RestExceptions from RestClient.

Includes test code in PullBase to force exceptions to test new logic. SuspendQueueException may
not be the proper Queue Exception to throw - it seems to delete/‘release’ the problematic item when
it should be kept so it can be processed fully.