12345678901234567890123456789012345678901234567890123456789012345678901234567890 Completes intelligent Drupal entity ID upsert code in PullBase
If the SF API cannot be reached while trying to save the entity ID in SF, the queue processing is
halted and the current queue item and remaining queue items are kept to process on a future pass.
A check has been added on PullBase::updateEntity() to send the Entity ID to SF if none exists and
the ID is configured to sync.
Adds PullException to facilitate throwing expected Exception up to Cron to tell Cron to suspend
queue processing but keep current and remaining queue items.